Added refactored nsdoc foundation
